Implementa regras de negócio

+++ SistemaReserva.Application/Recursos/DeleteRecurso/DeleteRecursoService.cs
@@ -1,4 +1,5 @@
-﻿using SistemReserva.Domain.Interfaces;
+﻿using SistemReserva.Domain.Exceptions;
+using SistemReserva.Domain.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -19,7 +20,12 @@ namespace SistemReserva.Application.Recursos.DeleteRecurso
             var recurso = await _unitOfWork.RecursoRepository.GetByIdAsync(id);
             if (recurso == null)
             {
-                throw new Exception("Recurso não encontrado");
+                throw new NotFoundException("Recurso não encontrado");
+            }
+            var possuiReservas = await _unitOfWork.RecursoRepository.PossuiReservasAsync(id);
+            if(possuiReservas)
+            {
+                throw new BadRequestException("Não é possível excluir um recurso com reservas vinculadas.");
             }
             await _unitOfWork.RecursoRepository.RemoveAsync(recurso);
             await _unitOfWork.CommitAync();
